+++ HISTORY.rst
@@ -4,6 +4,11 @@
 * Functions from the pthread library are now supported
 * Added functions: ``read(2)``
 * Added constants: ``NSIG``
+* Added new module :mod:`glibc_select` that contains a Python 2.7+ version of the
+  select.py from Python 3.4. This module contains a pure-python version of the
+  ``epoll`` class. It may be used in place of the module from the standard
+  library if additional features are desired in a cross-python-version portable
+  manner.
 
 0.5 (2014-10-22)
 ================